check now command now interacts with DB rather than monitor class specifically migrated check now command to FastAPI
